Day 1 deep dive in Linux
oin us for Day 1 of our deep dive into Linux as we prepare for DEVOPS. We'll explore the fundamentals of Linux, including its architecture, command line interface, and file system, as well as some essential tools and techniques that will help us to succeed in a DEVOPS environment. Whether you're a seasoned Linux user or just starting out, this video is a great way to get started on your journey towards becoming a successful DEVOPS professional.