Create Day 3 Linux Fundamentals, User, Groups, File Permission
Day 2 Linux Fundamentals, User, Groups, File Permission - Preparation for DEVOPS" is an essential video for anyone looking to learn more about Linux. In this video, you'll delve deeper into the operating system, covering topics such as user and group management, file permissions, and more. These concepts are critical for anyone interested in pursuing a career in DevOps, as they form the foundation of many DevOps practices and tools. With clear explanations and practical examples, this video is an excellent resource for anyone looking to strengthen their knowledge of Linux and prepare for a successful career in DevOps.