Create Day 6 Linux LVM Merge | Split | Migrate
Day 6Linux LVM (merge, split, migrate) - Preparation for DEVOPS" is an informative video that builds upon the concepts covered in previous videos to delve deeper into LVM. In this video, you'll learn how to perform advanced LVM operations such as merging and splitting logical volumes and migrating data between physical volumes. These skills are critical for DevOps professionals as they enable efficient use of storage resources and facilitate seamless infrastructure management. With practical examples and clear explanations, this video is an excellent resource for anyone looking to enhance their knowledge of LVM and prepare for a successful career in DevOps.