Create Day 8 Linux Create and Managing RAID  - Preparation for DEVOPS
Day 8 Linux Create and Managing RAID - Preparation for DEVOPS" is an essential video for DevOps professionals who want to learn about RAID (Redundant Array of Independent Disks) and how to create and manage RAID arrays on Linux systems. In this video, you'll learn about the various RAID levels and their characteristics, as well as how to create and manage RAID arrays using both software and hardware RAID controllers. You'll also learn how to monitor the health of RAID arrays and how to replace failed disks in the array to maintain data availability. With clear explanations and practical examples, this video is an excellent resource for anyone looking to enhance their knowledge of RAID and prepare for a successful career in DevOps.